refactor(feed): migrate feed to module layout
