+++ src/screens/feed/index.ts
@@ -1,0 +1,4 @@
+export { CreatePostScreen } from './view/CreatePostScreen';
+export { FeedScreen } from './view/FeedScreen';
+export { PostDetailsScreen } from './view/PostDetailsScreen';
+export * from './routes';
